moved addDatasetVersionCondensed into versions.go, swagger specs updated
